Paril/geometry shaders (#483)
* Geooooometry

* Shader cleanup
Remove the random offset in lightning trail. It's not looking very good, even with attempts to keep the vertices connected.
Add cvar for geometry shaders and their control, fall back to 0 if they can't be loaded

* Fix splash particles

* Use slightly more efficient caustic code. Looks about the same. Should prevent bubbles on Pan's hardware hopefully.

* Fix Autotools build.

* Fix Xcode build.

* Fix a buffer overreach warning.

* Fix several warnings around lossy implicit conversion.

* Fix shadows, hopefully?

* Remove "size", auto-calculate as well.
Revert 590237599c030a15b0f35f66cc1ed105dd37c358, sky issues resolved for now. Sadface.

* Fix up the construction of buffers with the concept of jdolan's struct majigger.
Remove the requirement of setting a "count" if the count is 1.
Rename some stuff to make more sense; R_ATTRIB_x is now R_TYPE_x, and R_ARRAY_x is now R_ATTRIB_x.
Clean up pc-windows copy procedure; use an alias, set retry time to 3s instead of 30s, redirect output to nul, and fix up double quoting.

+++ src/game/default/g_map_list.c
@@ -275,7 +275,7 @@ const g_map_list_map_t *G_MapList_Next(void) {
 
 	GList *list = NULL;
 
-	const size_t len = g_list_length(g_map_list);
+	const guint len = g_list_length(g_map_list);
 	if (len) {
 		if (g_random_map->value) {
 			list = g_list_nth(g_map_list, Randomr(0, len));
